Split test_client.py instead of bumping timeout.
