+++ python/ray/tests/BUILD
@@ -53,6 +53,8 @@ py_test_module_list(
 
 py_test_module_list(
   files = [
+    "test_client_1.py",
+    "test_client_2.py",
     "test_client_builder.py",
     "test_client_compat.py",
     "test_client_init.py",
@@ -65,17 +67,7 @@ py_test_module_list(
   ],
   size = "medium",
   extra_srcs = SRCS,
-  tags = ["exclusive", "client_tests", "medium_client_tests", "team:serverless"],
-  deps = ["//:ray_lib"],
-)
-
-py_test_module_list(
-  files = [
-    "test_client.py",
-  ],
-  size = "large",
-  extra_srcs = SRCS,
-  tags = ["exclusive", "client_tests", "large_client_tests", "team:serverless"],
+  tags = ["exclusive", "client_tests", "team:serverless"],
   deps = ["//:ray_lib"],
 )
